Un-break things from previous commit

+++ src/eckit/container/Trie.cc
@@ -69,9 +69,9 @@ bool Trie<T>::remove(const unsigned char* key)
         return (kids_.size() == 0); // Remove me if size if 0
 	}
 
-    size_t pos = int(*key) - int(from_);
+    int pos = int(*key) - int(from_);
 
-    if( pos >= 0 && pos  < kids_.size() && kids_[pos])
+    if( pos >= 0 && pos  < int(kids_.size()) && kids_[pos])
         if(kids_[pos]->remove(key+1))
         {
             delete kids_[pos];
@@ -97,15 +97,15 @@ Trie<T>* Trie<T>::find(const unsigned char* key, bool make)
 	if(*key == 0) 
 		return this;
 
-    size_t pos = int(*key) - int(from_);
+    int pos = int(*key) - int(from_);
 
-    if( pos >= 0 && pos  < kids_.size() && kids_[pos])
+    if( pos >= 0 && pos  < int(kids_.size()) && kids_[pos])
         return kids_[pos]->find(key+1, make);
 
 	if(make) {
 
         // If outside of range, we need to expand the storage vector.
-        if(pos < 0 || pos >= kids_.size()) {
+        if(pos < 0 || pos >= int(kids_.size())) {
 
 			int from = 0;
